+++ AlchemistCode/Cards/Uncommon/Slake.cs
@@ -1,0 +1,23 @@
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using Alchemist.AlchemistCode.Powers;
+
+namespace Alchemist.AlchemistCode.Cards.Uncommon;
+
+// The Antitoxin generator that also does something on the turn you play it. Priced off Iron Wave, a
+// Common at 5 damage plus 5 Block
+public class Slake : AlchemistCard
+{
+    public Slake() : base(1, CardType.Attack, CardRarity.Uncommon, TargetType.AnyEnemy)
+    {
+        WithDamage(8, 3);
+        WithPower<AntitoxinPower>(3, 1);
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await CommonActions.CardAttack(this, play, vfx: HitVfx("vfx/vfx_bloody_impact")).Execute(choiceContext);
+        await CommonActions.ApplySelf<AntitoxinPower>(choiceContext, this);
+    }
+}